+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### 🚨 Breaking changes
 
-- [](https://github.com/vegaprotocol/vega/issues/xxxx) -
+- [9945](https://github.com/vegaprotocol/vega/issues/9945) - Add liquidation strategy.
 
 ### 🗑️ Deprecation
 
@@ -18,6 +18,8 @@
 - [9955](https://github.com/vegaprotocol/vega/issues/9955) - Add data node subscription for transaction results.
 - [10004](https://github.com/vegaprotocol/vega/issues/10004) Track average entry price in position engine
 - [9825](https://github.com/vegaprotocol/vega/issues/9825) - Remove quadratic slippage.
+- [9516](https://github.com/vegaprotocol/vega/issues/9516) - Add filter by transfer ID for ledger entries API.
+- [9943](https://github.com/vegaprotocol/vega/issues/9943) - Support amending the order size by defining the target size.
 
 ### 🐛 Fixes